Separate types package
Used by api
