+++ apps/api/Dockerfile
@@ -1,14 +1,8 @@
-FROM node:18.12.0-alpine as builder
-WORKDIR /app
-COPY package*.json ./
-COPY . .
-RUN yarn install && yarn run build
-
 FROM node:18.12.0-alpine
 WORKDIR /app
 COPY package*.json ./
 RUN yarn install --only=production && yarn cache clean
-COPY --from=builder /app/build ./build
+COPY build ./build
 EXPOSE 80
 ARG version="none"
 ENV NODE_ENV=production
