+++ apps/api/package.json
@@ -41,7 +41,8 @@
     "eslint": "8.26.0",
     "prettier": "2.7.1",
     "ts-node": "10.9.1",
-    "ts-node-dev": "2.0.0"
+    "ts-node-dev": "2.0.0",
+    "types": "*"
   },
   "packageManager": "yarn@3.2.4"
 }
