+++ apps/api/src/models/data/description.ts
@@ -1,24 +1,7 @@
 import { Knex } from "knex";
 import db from "../../db";
 import { Filter } from ".";
-
-export interface Description {
-  id: string;
-  dg_id: string | null;
-  include: number | null;
-  title: string | null;
-  name: string | null;
-  type: string | null;
-  min_denominator: number | null;
-  level_green: number | null;
-  level_yellow: number | null;
-  level_direction: number | null;
-  short_description: string | null;
-  long_description: string | null;
-  registry_id: number;
-  rname: string | null;
-  full_name: string;
-}
+import { Description } from "types";
 
 export const descriptionModel = (
   register: string,
