+++ apps/api/src/models/info/names.ts
@@ -1,15 +1,5 @@
 import db from "../../db";
-
-export interface RegisterName {
-  id: number;
-  rname: string;
-  full_name: string;
-  caregiver_data: 0 | 1 | null;
-  resident_data: 0 | 1 | null;
-  dg_data: 0 | 1 | null;
-  url: string | null;
-  description: string | null;
-}
+import { RegisterName } from "types";
 
 export const registerNamesModel = (): Promise<RegisterName[]> =>
   db
