+++ packages/types/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "types",
+  "version": "0.0.0",
+  "main": "./src/index",
+  "types": "./src/index",
+  "packageManager": "yarn@3.2.4",
+  "private": true,
+  "devDependencies": {
+    "typescript": "^4.8.4"
+  }
+}
