+++ packages/types/tsconfig.json
@@ -1,0 +1,9 @@
+{
+    "compilerOptions": {
+      "target": "es5",
+      "lib": ["dom", "dom.iterable", "esnext"],
+      "allowJs": true,
+      "strict": true
+    },
+    "include": ["src"]
+}
